home_page test and layout refactor
